requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
python-multipart
python-dotenv
email-validator